First Commit
This provides the basic foundation of the program

Notes: things to think about for future changes
cards.xml: the xml file to parse
Card.java: the Card object, all MTG cards are cards
Creature.java: extends Card. Only MTG creature are creatures
Planeswalker.java: extends Card. Only MTG planeswalkers are
planeswalkes.
XMLParser: This takes in an XML file of the format specified (cards.xml
as an example), and stores the data in card objects,.
testClass: a utility class allows me to test functionality of the
program